Huge reorganizing and interfacing commit
Driver and rider registration and data storage is now entirely handled by the registration contract. The driver and rider.sol interfacing contracts now interface with this contract in order to register and access registration data. Smart contracts need to be linked up when deploying in order to use each other and give each other permission to use each other!